AIOT-39: Add a log handler for mt8183
Currently, the only way to get traces from mt8183 is to use a shared
buffer, that could be use by mt8183 to write traces, and read by the
kernel to print them.

To implement and use it, we have redefined some functions, such as
printf. This was working but this not reusable, not reliable
(when linking with other libraries implementing them).

Instead, implements a log handler that will be called by metal_log
function. In that way, we juste have to init the log handler and
use metal_log to print the traces and make them available to the
kernel.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>